+++ src/assembly/AssemblyTwoAdress.java
@@ -15,19 +15,19 @@ public class AssemblyTwoAdress extends Assembly{
 
     public static void add(String op1,String op2)
     {
-        MyDesktop.getMemory().store(op1,MyDesktop.getMemory().retrieve(op1)+MyDesktop.getMemory().retrieve(op2));
+        MyDesktop.getMemory().store(op1,(Integer)MyDesktop.getMemory().retrieve(op1)+(Integer)MyDesktop.getMemory().retrieve(op2));
     }
     public static void sub(String op1,String op2)
     {
-        MyDesktop.getMemory().store(op1,MyDesktop.getMemory().retrieve(op1)-MyDesktop.getMemory().retrieve(op2));
+        MyDesktop.getMemory().store(op1,(Integer)MyDesktop.getMemory().retrieve(op1)-(Integer)MyDesktop.getMemory().retrieve(op2));
     }
     public static void mpy(String op1,String op2)
     {
-        MyDesktop.getMemory().store(op1,MyDesktop.getMemory().retrieve(op1)*MyDesktop.getMemory().retrieve(op2));
+        MyDesktop.getMemory().store(op1,(Integer)MyDesktop.getMemory().retrieve(op1)*(Integer)MyDesktop.getMemory().retrieve(op2));
     }
     public static void div(String op1,String op2)
     {
-        MyDesktop.getMemory().store(op1,MyDesktop.getMemory().retrieve(op1)/MyDesktop.getMemory().retrieve(op2));
+        MyDesktop.getMemory().store(op1,(Integer)MyDesktop.getMemory().retrieve(op1)/(Integer)MyDesktop.getMemory().retrieve(op2));
     }
     
 }
